chore: Adjust for internal member
